infra: reduce runner idle timeout to 15min, fix int->float parsing (#540)
## Summary
- Reduce idle timeout from 1h to 15min (0.25h) — solo dev project, no
need for long idle periods
- Fix `int()` -> `float()` for `IDLE_TIMEOUT_HOURS` env var parsing
(crashed on "0.25")

## Test plan
- [x] Pulumi applied locally, function deployed via gcloud
- [x] All 7 runners successfully stopped (verified TERMINATED status)
- [x] Logs confirm: creationTimestamp fallback, parallel stops, all
runners managed

@coderabbitai ignore

+++ infrastructure/functions/runner_manager/main.py
@@ -34,7 +34,7 @@ PROJECT_ID = os.environ.get("GCP_PROJECT", "nomadkaraoke")
 ZONE = os.environ.get("GCP_ZONE", "us-central1-a")
 WEBHOOK_SECRET_NAME = os.environ.get("WEBHOOK_SECRET_NAME", "github-webhook-secret")
 RUNNER_PAT_SECRET_NAME = os.environ.get("RUNNER_PAT_SECRET_NAME", "github-runner-pat")
-IDLE_TIMEOUT_HOURS = int(os.environ.get("IDLE_TIMEOUT_HOURS", "1"))
+IDLE_TIMEOUT_HOURS = float(os.environ.get("IDLE_TIMEOUT_HOURS", "1"))
 RUNNER_NAMES = os.environ.get(
     "RUNNER_NAMES",
     "github-runner-1,github-runner-2,github-runner-3,github-build-runner",
